Docker_all_model/requirements.txt
Flask
tensorflow
numpy
googletrans==4.0.0-rc1
langdetect
python-dotenv
